Move native-finder's arch-specific code to arch files
