+++ plugins/inputs/procstat/arch_ohers.go
@@ -1,0 +1,9 @@
+//go:build !linux && !windows
+
+package procstat
+
+import "github.com/shirou/gopsutil/v3/process"
+
+func processName(p *process.Process) (string, error) {
+	return p.Exe()
+}
